Backend/requirements.txt
acres==0.5.0
ag-ui-protocol==0.1.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
ansicon==1.89.0
anthropic==0.68.0
anyio==4.11.0
arabic-reshaper==3.0.0
argcomplete==3.6.2
arrow==1.3.0
asgiref==3.9.2
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.3.0
blessed==1.22.0
boto3==1.40.39
botocore==1.40.39
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
ci-info==0.3.0
click==8.3.0
cohere==5.18.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.1
cssselect2==0.8.0
distro==1.9.0
Django==5.2.6
django-appconf==1.1.0
django-cors-headers==4.9.0
django-cryptography==1.1
django-encrypted-model-fields==0.6.5
django-environ==0.12.1
django-filter==25.2
django-guardian==3.2.0
django-picklefield==3.3
django-q2==1.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docstring_parser==0.17.0
docx==0.2.4
duckdb==1.4.4
et_xmlfile==2.0.0
etelemetry==0.3.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.1
fastavro==1.12.0
filelock==3.19.1
freetype-py==2.5.1
frozenlist==1.7.0
fsspec==2025.9.0
genai-prices==0.0.27
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.40.3
google-auth-httplib2==0.3.0
google-genai==1.38.0
googleapis-common-protos==1.70.0
griffe==1.14.0
groq==0.31.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
invoke==2.2.0
isodate==0.7.2
jinxed==1.3.0
jiter==0.11.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwt==1.4.0
langchain-core==1.2.17
langchain-openai==1.1.10
langsmith==0.7.14
logfire==4.10.0
logfire-api==4.10.0
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.15.0
mdurl==0.1.2
mistralai==1.9.10
multidict==6.6.4
networkx==3.4.2
nexus-rpc==1.1.0
nibabel==5.3.3
nipype==1.10.0
numpy==2.2.6
openai==2.26.0
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-httpx==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.11.7
oscrypto==1.3.0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psycopg2-binary==2.9.11
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.11.9
pydantic-ai==1.0.10
pydantic-ai-slim==1.0.10
pydantic-evals==1.0.10
pydantic-graph==1.0.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
PyMuPDF==1.26.7
pyparsing==3.2.5
pypdf==6.6.0
pyperclip==1.10.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-http-client==3.3.7
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pyxnat==1.6.4
PyYAML==6.0.2
rdflib==7.5.0
redis==3.5.3
referencing==0.36.2
regex==2026.2.28
reportlab==4.4.6
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rlPyCairo==0.4.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
scipy==1.15.3
sendgrid==6.12.5
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.48.0
stripe==14.3.0
svglib==1.6.0
temporalio==1.17.0
tenacity==9.1.2
tiktoken==0.12.0
tinycss2==1.5.1
tokenizers==0.22.1
tomli==2.2.1
tqdm==4.67.1
traits==7.0.2
types-protobuf==6.32.1.20250918
types-python-dateutil==2.9.0.20251008
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.5.0
uuid_utils==0.14.1
uvicorn==0.37.0
wcwidth==0.2.14
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.4
whitenoise==6.11.0
wkhtmltopdf==0.2
wrapt==1.17.3
xhtml2pdf==0.2.17
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
